Generate crossword constraint problem on commandline
This commit replaces the DIMACS parsing in glucose with logic to
synthesize CNF clauses from a crossword problem. The wordlist and grid
(in plain ASCII format) are command line arguments.

The solver uses sophisticated techniques to try to narrow the search
space, for faster solving. In addition, it constrains the letter
choices of "distant" cells to popular letters, as a command line
option.